Population preparation done & dispatchers seem to run (Single, Feedback, Hungarian)

+++ playgrounds/ivtExt/src/main/java/playground/sebhoerl/avtaxi/framework/AVConfigGroup.java
@@ -28,7 +28,7 @@ public class AVConfigGroup extends ReflectiveConfigGroup {
 	}
 
 	@StringGetter(PARALLEL_ROUTERS)
-	public double getParallelRouters() {
+	public long getParallelRouters() {
 		return parallelRouters;
 	}
